backend/services/storage/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1
qdrant-client>=1.11.0
openai>=1.3.7
sentence-transformers>=2.2.2
PyPDF2>=3.0.1
python-docx>=1.1.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.1